Made constants and handler packages internal in the gerdana binding
