+++ addons/binding/org.openhab.binding.gardena/META-INF/MANIFEST.MF
@@ -8,9 +8,6 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-SymbolicName: org.openhab.binding.gardena;singleton:=true
 Bundle-Vendor: openHAB
 Bundle-Version: 2.4.0.qualifier
-Export-Package: 
- org.openhab.binding.gardena,
- org.openhab.binding.gardena.handler
 Import-Package: 
  com.google.common.collect,
  com.google.gson,
@@ -35,8 +32,6 @@ Import-Package:
  org.eclipse.smarthome.core.thing.binding.builder,
  org.eclipse.smarthome.core.thing.type,
  org.eclipse.smarthome.core.types,
- org.openhab.binding.gardena,
- org.openhab.binding.gardena.handler,
  org.osgi.framework,
  org.slf4j
 Service-Component: OSGI-INF/*.xml
