+++ addons/binding/org.openhab.binding.gardena/src/main/java/org/openhab/binding/gardena/internal/handler/AccountHandlerNotAvailableException.java
@@ -6,14 +6,19 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.gardena.handler;
+package org.openhab.binding.gardena.internal.handler;
 
 /**
- * The {@link GardenaDeviceConfig} class represents the configuration for a device connected to an Gardena account.
+ * Exception if the AccountHandler is not available.
  *
  * @author Gerhard Riegler - Initial contribution
  */
+public class AccountHandlerNotAvailableException extends Exception {
 
-public class GardenaDeviceConfig {
-    public String deviceId;
+    private static final long serialVersionUID = -1895774551653276530L;
+
+    public AccountHandlerNotAvailableException(String message) {
+        super(message);
+    }
+
 }
