+++ addons/binding/org.openhab.binding.gardena/src/main/java/org/openhab/binding/gardena/internal/handler/GardenaDeviceConfig.java
@@ -6,19 +6,14 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.gardena.handler;
+package org.openhab.binding.gardena.internal.handler;
 
 /**
- * Exception if the AccountHandler is not available.
+ * The {@link GardenaDeviceConfig} class represents the configuration for a device connected to an Gardena account.
  *
  * @author Gerhard Riegler - Initial contribution
  */
-public class AccountHandlerNotAvailableException extends Exception {
 
-    private static final long serialVersionUID = -1895774551653276530L;
-
-    public AccountHandlerNotAvailableException(String message) {
-        super(message);
-    }
-
+public class GardenaDeviceConfig {
+    public String deviceId;
 }
